+++ Abs/Entities/EnemyBullet.cpp
@@ -1,6 +1,6 @@
 #include "EnemyBullet.h"
 
-Abs::EnemyBullet::EnemyBullet(float x, float y, float w, float h) : EnemyEntity(x, y, w, h){
+Abs::EnemyBullet::EnemyBullet(double x, double y, double w, double h) : EnemyEntity(x, y, w, h){
 
 }
