+++ Abs/Entities/PosBonus.cpp
@@ -1,5 +1,5 @@
 #include "PosBonus.h"
 
-Abs::PosBonus::PosBonus(float x, float y, float w, float h) : BonusEntity(POSBONUS, x, y, w, h){
+Abs::PosBonus::PosBonus(double x, double y, double w, double h) : BonusEntity(POSBONUS, x, y, w, h){
 
 }
