+++ SDL/Entities/SDLEnemyShip.cpp
@@ -1,6 +1,6 @@
 #include "SDLEnemyShip.h"
 
-Sdl::SDLEnemyShip::SDLEnemyShip(Sdl::SDLWindow* window, float x, float y, float w, float h, ENTITY e) : EnemyShip(x, y, w, h, e){
+Sdl::SDLEnemyShip::SDLEnemyShip(Sdl::SDLWindow* window, double x, double y, double w, double h, ENTITY e) : EnemyShip(x, y, w, h, e){
     this->window = window;
 }
